requirements.txt
# Data Ingestion
requests==2.31.0
fredapi==0.5.1
pycoingecko==3.1.0
yfinance>=0.2.40

# Data Processing
pandas>=2.2.0
numpy>=1.26.0
apache-airflow==2.7.0

# Analytics
scipy==1.11.2
statsmodels==0.14.0
scikit-learn==1.3.0

# Visualization
plotly==5.16.1
dash==2.13.0
streamlit>=1.39.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7


# Cloud
boto3>=1.34.0
awscli==1.29.25

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1






